Fixes initial frame issue
The first time the Message Banner is shown it slides in from an arbitrary position at the top left. As far as I can tell this is due to the constraints being applied outside of the animation block.

Also setting `currentMessageBanner.center = target` is really not needed as this is using constraints which override setting the center anyway. I would drop setting `currentMessageBanner.center = target`  as well as `calculateTargetCenter:` altogether except that there are side effects in `calculateTargetCenter:` which are needed.