Fix for iOS 9 Banner not taking navigation bars into account properly
This appears to be a bug in iOS 9?
The statement `realtViewController.topLayoutGuide` seems to go away at some point after we set it here. To be clear it is not `nil` at run time when `[... constraintWithItem:realViewController.topLayoutGuide ...]` is called however when I bring this up in Reveal App the constraint is labeled as `???.bottom = MBLMessageBannerView:0x...` and the `First Item` is `Unknown`.

After a few hours banging my head against a wall this was the best I could come up with and it seems to work but I am skeptical as to how robust this really is.